requirement.txt
mysqlclient = "==1.4.6"
djangorestframework = "==3.11.2"
django-filter = "==2.2.0"
channels = "==2.4.0"
channels-redis = "==2.4.2"
pytz = "*"
Django = "==3.0.7"
Markdown = "==3.2.1"